requirements.txt
langgraph>=0.2.0
pydantic>=2.0.0
google-generativeai>=0.8.0
tavily-python>=0.3.0
pytest>=8.0.0
pytest-asyncio>=0.21.0
ruff>=0.5.0
black>=24.0.0
streamlit>=1.28.0
plotly>=5.17.0
python-dotenv>=1.0.0
